Call fetchAudioAndMarks to get the audio from the TTS api
